ponyracer/src/app/races-resolver.service.ts
import { Injectable } from '@angular/core';
import { ActivatedRouteSnapshot, Resolve, RouterStateSnapshot } from '@angular/router';
import { Observable } from 'rxjs/Observable';

import { RaceModel, RaceStatus } from './models/race.model';
import { RaceService } from './race.service';

@Injectable()
export class RacesResolverService implements Resolve<Array<RaceModel>> {

  constructor(private raceService: RaceService) { }

  resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): RaceModel[] | Observable<RaceModel[]> | Promise<RaceModel[]> {
    return this.raceService.list(route.routeConfig.path.toUpperCase() as RaceStatus);
  }

}
